refactor(shared): move feature-flag util into FSD shared layer
isFeatureEnabled + useFeatureFlag move from app/lib + app/hooks to
src/shared/lib/feature-flags (with an index barrel). The analysis slice's
ViewAnalysisTrigger now imports the hook from @/src/shared/lib/feature-flags
(downward features -> shared) instead of @/app/hooks (upward features -> app),
fixing an FSD direction violation. VectorAreasLayer repointed too. No behaviour
change.
