+++ src/shared/lib/feature-flags/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Public API of the shared feature-flag utility. Hidden features opt in via a
+ * single comma-separated URL param, e.g. `?ff=analysis`.
+ */
+export { isFeatureEnabled } from "./feature-flags";
+export { useFeatureFlag } from "./use-feature-flag";
